Moving DN files to DN package
