+++ app/src/main/java/io/plaidapp/designernews/DesignerNewsPrefs.java
@@ -1,0 +1,90 @@
+/*
+ *   Copyright 2018 Google LLC
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+
+package io.plaidapp.designernews;
+
+import android.content.Context;
+import android.support.annotation.NonNull;
+import android.text.TextUtils;
+
+import io.plaidapp.designernews.data.api.DesignerNewsService;
+import io.plaidapp.designernews.data.api.model.User;
+import io.plaidapp.designernews.login.data.DesignerNewsLoginRepository;
+import io.plaidapp.util.ShortcutHelper;
+
+/**
+ * Storing Designer News user state
+ */
+public class DesignerNewsPrefs {
+
+    private static volatile DesignerNewsPrefs singleton;
+
+    private DesignerNewsService api;
+    private DesignerNewsLoginRepository loginRepository;
+
+    public static DesignerNewsPrefs get(Context context) {
+        if (singleton == null) {
+            synchronized (DesignerNewsPrefs.class) {
+                singleton = new DesignerNewsPrefs(
+                        Injection.provideDesignerNewsLoginRepository(context));
+            }
+        }
+        return singleton;
+    }
+
+    private DesignerNewsPrefs(DesignerNewsLoginRepository loginRepository) {
+        this.loginRepository = loginRepository;
+    }
+
+    public boolean isLoggedIn() {
+        return loginRepository.isLoggedIn();
+    }
+
+    public void setAccessToken(@NonNull Context context, String accessToken) {
+        if (!TextUtils.isEmpty(accessToken)) {
+            loginRepository.setAccessToken(accessToken);
+            createApi();
+            ShortcutHelper.enablePostShortcut(context);
+        }
+    }
+
+    public void setLoggedInUser(User user) {
+        loginRepository.setLoggedInUser(user);
+    }
+
+    public User getUser() {
+        return loginRepository.getUser();
+    }
+
+    public void logout(@NonNull Context context) {
+        loginRepository.logout();
+        createApi();
+        ShortcutHelper.disablePostShortcut(context);
+    }
+
+    public DesignerNewsService getApi() {
+        if (api == null) {
+            createApi();
+        }
+        return api;
+    }
+
+    private void createApi() {
+        api = Injection.provideDesignerNewsService(loginRepository.getAccessToken());
+    }
+
+}
