+++ app/src/main/java/io/plaidapp/designernews/data/api/model/AccessToken.java
@@ -15,7 +15,7 @@
  *
  */
 
-package io.plaidapp.data.api.designernews.model;
+package io.plaidapp.designernews.data.api.model;
 
 /**
  * Models a Designer News API access token
