+++ app/src/main/java/io/plaidapp/designernews/data/api/model/Story.java
@@ -15,7 +15,7 @@
  *
  */
 
-package io.plaidapp.data.api.designernews.model;
+package io.plaidapp.designernews.data.api.model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
