+++ app/src/main/java/io/plaidapp/designernews/data/api/model/StoryLinks.kt
@@ -1,4 +1,4 @@
-package io.plaidapp.data.api.designernews.model
+package io.plaidapp.designernews.data.api.model
 
 import android.os.Parcel
 import android.os.Parcelable
