+++ app/src/main/java/io/plaidapp/ui/FeedAdapter.java
@@ -71,8 +71,8 @@ import io.plaidapp.R;
 import io.plaidapp.data.DataLoadingSubject;
 import io.plaidapp.data.PlaidItem;
 import io.plaidapp.data.PlaidItemSorting;
-import io.plaidapp.data.api.designernews.StoryWeigher;
-import io.plaidapp.data.api.designernews.model.Story;
+import io.plaidapp.designernews.data.api.StoryWeigher;
+import io.plaidapp.designernews.data.api.model.Story;
 import io.plaidapp.data.api.dribbble.PlayerShotsDataManager;
 import io.plaidapp.data.api.dribbble.ShotWeigher;
 import io.plaidapp.data.api.dribbble.model.Shot;
@@ -81,7 +81,6 @@ import io.plaidapp.data.api.producthunt.model.Post;
 import io.plaidapp.data.pocket.PocketUtils;
 import io.plaidapp.data.prefs.SourceManager;
 import io.plaidapp.ui.recyclerview.Divided;
-import io.plaidapp.ui.transitions.ReflowText;
 import io.plaidapp.ui.widget.BadgedFourThreeImageView;
 import io.plaidapp.ui.widget.BaselineGridTextView;
 import io.plaidapp.util.Activities;
