+++ app/src/main/java/io/plaidapp/ui/HomeActivity.java
@@ -82,11 +82,11 @@ import io.plaidapp.R;
 import io.plaidapp.data.DataManager;
 import io.plaidapp.data.PlaidItem;
 import io.plaidapp.data.Source;
-import io.plaidapp.data.api.designernews.PostStoryService;
-import io.plaidapp.data.api.designernews.model.Story;
+import io.plaidapp.designernews.data.api.PostStoryService;
+import io.plaidapp.designernews.data.api.model.Story;
 import io.plaidapp.data.api.dribbble.model.Shot;
 import io.plaidapp.data.pocket.PocketUtils;
-import io.plaidapp.data.prefs.DesignerNewsPrefs;
+import io.plaidapp.designernews.DesignerNewsPrefs;
 import io.plaidapp.data.prefs.DribbblePrefs;
 import io.plaidapp.data.prefs.SourceManager;
 import io.plaidapp.ui.recyclerview.FilterTouchHelperCallback;
