+++ features/activities/src/main/AndroidManifest.xml
@@ -64,7 +64,7 @@
             android:theme="@style/Plaid.Translucent.Dribbble.PlayerSheet" />
 
         <activity
-            android:name="io.plaidapp.ui.DesignerNewsStory"
+            android:name="io.plaidapp.ui.designernews.DesignerNewsStory"
             android:parentActivityName="io.plaidapp.ui.HomeActivity"
             android:theme="@style/Plaid.Translucent.DesignerNewsStory" />
 
@@ -100,7 +100,7 @@
         </activity>
 
         <activity
-            android:name="io.plaidapp.ui.DesignerNewsLogin"
+            android:name="io.plaidapp.ui.designernews.DesignerNewsLogin"
             android:exported="false"
             android:theme="@style/Plaid.Translucent.DesignerNewsLogin"
             android:windowSoftInputMode="adjustResize" />
@@ -111,7 +111,7 @@
             android:parentActivityName="io.plaidapp.ui.HomeActivity" />
 
         <activity
-            android:name="io.plaidapp.ui.PostNewDesignerNewsStory"
+            android:name="io.plaidapp.ui.designernews.PostNewDesignerNewsStory"
             android:exported="true"
             android:theme="@style/Plaid.Translucent.PostDesignerNewsStory">
             <intent-filter>
@@ -123,11 +123,11 @@
         </activity>
 
         <service
-            android:name="io.plaidapp.data.api.designernews.PostStoryService"
+            android:name="io.plaidapp.designernews.data.api.PostStoryService"
             android:exported="false" />
 
         <service
-            android:name="io.plaidapp.data.api.designernews.UpvoteStoryService"
+            android:name="io.plaidapp.designernews.data.api.UpvoteStoryService"
             android:exported="false" />
 
         <provider
