+++ features/activities/src/main/res/layout/activity_post_new_designer_news_story.xml
@@ -23,7 +23,7 @@
     android:id="@+id/bottom_sheet"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="io.plaidapp.ui.PostNewDesignerNewsStory">
+    tools:context="io.plaidapp.ui.designernews.PostNewDesignerNewsStory">
 
     <LinearLayout
         android:id="@+id/bottom_sheet_content"
